move device connect state enum to its own file
