+++ src/electron/device-connect-view/components/device-connect-state.tsx
@@ -1,0 +1,6 @@
+export enum DeviceConnectState {
+    Default,
+    Connecting,
+    Connected,
+    Error,
+}
